requirements.txt
Django==4.2.21
djangorestframework==3.14.0
beautifulsoup4==4.12.3
requests==2.31.0
python-dotenv==1.0.1
lxml==5.1.0
aiohttp==3.9.3
pytest==8.0.2
pytest-django==4.8.0
psycopg2-binary==2.9.9 